requirements.txt
pip-requirements
Flask==3.0.3
flask-cors==5.0.0
py-ballisticcalc[exts]
py-ballisticcalc[charts]
numpy==2.1.0
scipy==1.14.0
pyserial

